// yameza-be/src/modules/production/enums/production.enum.ts
export enum ProductionStatus {
  PENDING = 'PENDING',
  IN_PROGRESS = 'IN_PROGRESS',
  PAUSED = 'PAUSED',
  COMPLETED = 'COMPLETED',
}
export enum ProductionStageName {
  CORTE = 'CORTE',
  CANTEADO = 'CANTEADO',
  ABISAGRADO = 'ABISAGRADO',
  ENSAMBLADO = 'ENSAMBLADO',
  ACABADO = 'ACABADO',
}
export enum ProductionStageStatus {
  PENDING = 'PENDING',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  PAUSED = 'PAUSED',
}